requirements.txt
mysql-connector-python==8.0.32
matplotlib==3.7.1
python-dotenv>=1.0.0